moving post method from controller to model, adding response validations
