+++ lib/pagseguro.rb
@@ -9,6 +9,7 @@ require "pagseguro/faker"
 require "pagseguro/rake"
 require "pagseguro/railtie"
 require "pagseguro/notification"
+require "pagseguro/validator"
 require "pagseguro/order"
 require "pagseguro/action_controller"
 require "pagseguro/helper"
