+++ spec/pagseguro/rake_spec.rb
@@ -36,6 +36,7 @@ describe PagSeguro::Rake do
   end
 
   it "should ping return URL with default arguments" do
+    p params
     params["StatusTransacao"].should == "Completo"
     params["TipoPagamento"].should == "Cartão de Crédito"
   end
@@ -139,3 +140,4 @@ describe PagSeguro::Rake do
     $HTTP_PARAMS
   end
 end
+
